+++ src/com/upc/edu/pe/expose/DocumentoPago.java
@@ -1,9 +1,4 @@
-package com.upc.edu.pe.negocio;
-
-import com.upc.edu.pe.entidad.Documento;
-import com.upc.edu.pe.entidad.Factura;
-import com.upc.edu.pe.entidad.Persona;
-import sun.swing.BakedArrayList;
+package com.upc.edu.pe.expose;
 
 import java.util.ArrayList;
